requirements.txt
flask
flask-cors
invoke
pytest==7.4.3
pytest-flask==1.3.0
gtts
pandas
boto3
fastapi
comps
opea-comps
streamlit
boto3>=1.34.0
youtube-transcript-api>=0.6.2
chromadb>=0.4.22
watchdog==6.0.0
pydub
duckduckgo_search
openai
html2text
streamlit>=1.25.0
openai>=1.0.0
duckduckgo-search>=3.0.0
requests>=2.28.0
html2text>=2020.1.16
python-dotenv>=1.0.0
pandas>=1.5.0
pillow
